requirements.txt
# Core
pyspark==4.1.1
google-cloud-storage==2.19.0
google-cloud-bigquery==3.28.0
pandas==2.2.3
pyarrow==18.1.0

# Dashboard
streamlit==1.41.1
plotly==6.0.0
db-dtypes==1.3.1

# dbt
dbt-core==1.9.4
dbt-bigquery==1.9.2

# Utilities
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.67.1
